Refactored groups related methods from Util
