+++ src/main/java/net/sf/jabref/groups/GroupDialog.java
@@ -26,12 +26,12 @@ import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.fieldeditors.TextField;
 import net.sf.jabref.gui.keyboard.KeyBinds;
 import net.sf.jabref.gui.util.PositionWindow;
+import net.sf.jabref.logic.groups.GroupsUtil;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.search.SearchRules;
 import net.sf.jabref.logic.search.describer.SearchDescribers;
 import net.sf.jabref.logic.util.strings.StringUtil;
 import net.sf.jabref.model.entry.BibtexEntry;
-import net.sf.jabref.util.Util;
 
 import javax.swing.*;
 import javax.swing.event.CaretEvent;
@@ -494,7 +494,7 @@ class GroupDialog extends JDialog {
         if (!vec.isEmpty()) {
             BibtexEntry[] entries = new BibtexEntry[vec.size()];
             vec.toArray(entries);
-            if (!Util.warnAssignmentSideEffects(new AbstractGroup[] {m_resultingGroup},
+            if (!GroupsUtil.warnAssignmentSideEffects(new AbstractGroup[] {m_resultingGroup},
                     entries, m_basePanel.getDatabase(), this)) {
                 return;
             }
